+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- **Breaking:** `YoutubeDownloader\Provider\Youtube\Provider::createFromConfigAndToolkit()` was removed, use `YoutubeDownloader\Provider\Youtube\Provider::createFromOptions()` instead
+- **Breaking:** `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithConfig()` was removed, use `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithOptions()` instead
+- **Breaking:** `YoutubeDownloader\Toolkit::validateVideoId()` isn't used anymore and was removed
+- **Breaking:** `YoutubeDownloader\Toolkit::formatBytes()` isn't used anymore and was removed
+- **Breaking:** `YoutubeDownloader\Toolkit::is_chrome()` isn't used anymore and was removed
+- **Breaking:** `YoutubeDownloader\Toolkit::getDownloadMP3()` isn't used anymore and was removed
+
 ## [0.6] - 2018-01-02
 
 ### Added
@@ -32,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithConfig()` will be removed in 0.7, use `YoutubeDownloader\Provider\Youtube\VideoInfo::createFromStringWithOptions()` instead
 - `YoutubeDownloader\Toolkit::validateVideoId()` isn't used anymore and will be removed in 0.7
 - `YoutubeDownloader\Toolkit::formatBytes()` isn't used anymore and will be removed in 0.7
+- `YoutubeDownloader\Toolkit::is_chrome()` isn't used anymore and will be removed in 0.7
 - `YoutubeDownloader\Toolkit::getDownloadMP3()` isn't used anymore and will be removed in 0.7
 
 ### Removed